+++ tests/testthat/test-gh_token.R
@@ -65,7 +65,10 @@ test_that("validate_gh_pat() rejects bad characters, wrong # of characters", {
   # newer PATs
   expect_error(gh_pat(paste0("ghp_", strrep("B", 36))), NA)
   expect_error(gh_pat(paste0("ghp_", strrep("3", 251))), NA)
+  expect_error(gh_pat(paste0("github_pat_", strrep("A", 36))), NA)
+  expect_error(gh_pat(paste0("github_pat_", strrep("3", 244))), NA)
   expect_error(gh_pat(paste0("ghJ_", strrep("a", 36))), "prefix", class = "error")
+  expect_error(gh_pat(paste0("github_pa_", strrep("B", 244))), "github_pat_", class = "error")
 })
 
 test_that("format.gh_pat() and str.gh_pat() hide the middle stuff", {