+++ NEWS.md
@@ -1,5 +1,9 @@
 # gh (development version)
 
+* gh can now validate GitHub
+  [fine-grained](https://github.blog/2022-10-18-introducing-fine-grained-personal-access-tokens-for-github/)
+  personal access tokens (@jvstein, #171).
+
 # gh 1.3.1
 
 * gh now accepts lower-case methods i.e. both `gh::gh("get /users/hadley/repos")` and `gh::gh("GET /users/hadley/repos")` work (@maelle, #167).